core/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=core/requirements.txt core/requirements.in
#
aiopg==1.3.0
    # via -r core/requirements.in
asgiref==3.4.1
    # via uvicorn
async-timeout==3.0.1
    # via aiopg
certifi==2021.5.30
    # via -r core/requirements.in
click==7.1.2
    # via
    #   peewee-migrations
    #   typer
    #   uvicorn
fastapi==0.66.0
    # via -r core/requirements.in
funcy==1.16
    # via -r core/requirements.in
h11==0.12.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
peewee==3.14.4
    # via
    #   -r core/requirements.in
    #   peewee-async
    #   peewee-migrations
peewee-async==0.7.1
    # via -r core/requirements.in
peewee-migrations==0.3.19
    # via -r core/requirements.in
psycopg2-binary==2.9.1
    # via aiopg
pydantic==1.8.2
    # via fastapi
python-dotenv==0.18.0
    # via uvicorn
pyyaml==5.4.1
    # via uvicorn
starlette==0.14.2
    # via fastapi
typer==0.3.2
    # via -r core/requirements.in
typing-extensions==3.10.0.0
    # via pydantic
ujson==4.0.2
    # via -r core/requirements.in
uvicorn[standard]==0.14.0
    # via -r core/requirements.in
uvloop==0.15.2
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==9.1
    # via uvicorn
